(svn r1488) -Fix: fixed some spelling mistakes (tnx to Tron) -Fix: fixed some compiler-warnings

+++ ttd.c
@@ -657,7 +657,7 @@ int ttd_main(int argc, char* argv[])
 	if (sounddriver[0]) ttd_strlcpy(_ini_sounddriver, sounddriver, sizeof(_ini_sounddriver));
 	if (videodriver[0]) ttd_strlcpy(_ini_videodriver, videodriver, sizeof(_ini_videodriver));
 	if (resolution[0]) { _cur_resolution[0] = resolution[0]; _cur_resolution[1] = resolution[1]; }
-	if (startdate != -1) _patches.starting_date = startdate;
+	if (startdate != (uint)-1) _patches.starting_date = startdate;
 
 	if (_dedicated_forks && !_dedicated_enabled)
 		_dedicated_forks = false;
